Moved the addWebPartToPage static function from Helper.WebPart to the Helper.methods directory. Added a new function for adding content editor webparts.
