+++ @types/helper/webpart.d.ts
@@ -14,13 +14,6 @@ export interface IWebPart {
     new(props: IWebPartProps);
 
     /**
-     * Adds a script editor webpart to a page.
-     * @param url - The relative url of the page.
-     * @param wpProps - The webpart properties.
-     */
-    addWebPartToPage(url: string, wpProps: IWebPartScriptEditor): PromiseLike<void>;
-
-    /**
      * Creates an instance of a webpart.
      * @param props - The webpart properties.
      */
@@ -81,27 +74,4 @@ export interface IWebPartProps {
 
     /** The target element id to render the webpart to */
     elementId: string;
-}
-
-/**
- * The script editor webpart properties
- */
-export interface IWebPartScriptEditor {
-    /** The webpart description. */
-    description?: string;
-
-    /** The webpart chrome type. (BorderOnly, Default, None, TitleAndBorder or TitleOnly) */
-    chromeType?: string;
-
-    /** The webpart content. */
-    content: string;
-
-    /** The webpart index. */
-    index?: number;
-
-    /** The webpart title. */
-    title?: string;
-
-    /** The webpart zone. */
-    zone?: string;
 }
